Add byScreen tab to website
